hello

+++ database/migrations/2020_07_19_154452_create_chart_totalquantityandcropprice_view.php
@@ -16,11 +16,11 @@ class CreateChartTotalquantityandcroppriceView extends Migration
         \DB::statement("
         CREATE VIEW chart_totalquantityandcropprice
         AS
-        SELECT 
+        SELECT
         posts.crop_name AS crop_name,
-        SUM(posts.crop_quantity::decimal) AS sumcropqty,
-        (SUM(posts.crop_quantity::decimal) * posts.crop_price::decimal) AS allcropprice,
-        AVG(posts.crop_price::decimal) AS avgcropprice,
+        SUM(posts.crop_quantity) AS sumcropqty,
+        (SUM(posts.crop_quantity) * posts.crop_price) AS allcropprice,
+        AVG(posts.crop_price) AS avgcropprice,
         created_at
         FROM
         posts
@@ -35,6 +35,5 @@ class CreateChartTotalquantityandcroppriceView extends Migration
      */
     public function down()
     {
-
     }
 }
