+++ database/migrations/2020_07_19_154942_create_farmer_sales_view.php
@@ -15,13 +15,13 @@ class CreateFarmerSalesView extends Migration
     {
         \DB::statement("
         CREATE VIEW farmer_sales AS
-        SELECT 
+        SELECT
             posts.user_id AS user_id,
             posts.crop_name AS crop_name,
-            SUM(posts.fixed_quantity::decimal) AS totalfixedqty,
-            SUM(posts.crop_quantity::decimal) AS totalavailableqty,
-            SUM(posts.kilogram_sold::decimal) AS totalkgsold,
-            AVG(ROUND(posts.percentage_sold_before_harvest::decimal,
+            SUM(posts.fixed_quantity) AS totalfixedqty,
+            SUM(posts.crop_quantity) AS totalavailableqty,
+            SUM(posts.kilogram_sold) AS totalkgsold,
+            AVG(ROUND(posts.percentage_sold_before_harvest,
                     1)) AS totalpercentage
         FROM
             posts
@@ -36,6 +36,5 @@ class CreateFarmerSalesView extends Migration
      */
     public function down()
     {
-
     }
 }
