+++ database/migrations/2020_07_19_155204_create_farmer_totalqty_view.php
@@ -15,11 +15,11 @@ class CreateFarmerTotalqtyView extends Migration
     {
         \DB::statement("
         CREATE VIEW farmer_totalqty AS
-        SELECT 
+        SELECT
             posts.user_id AS user_id,
-            SUM(posts.crop_quantity::decimal) AS sumcropqty,
-            (SUM(posts.crop_quantity::decimal) * posts.crop_price::decimal) AS allcropprice,
-            AVG(posts.crop_price::decimal) AS avgcropprice,
+            SUM(posts.crop_quantity) AS sumcropqty,
+            (SUM(posts.crop_quantity) * posts.crop_price) AS allcropprice,
+            AVG(posts.crop_price) AS avgcropprice,
             posts.crop_name AS crop_name
         FROM
             posts
