+++ database/migrations/2020_07_19_161922_create_view_users_posts_up_ul_view.php
@@ -15,7 +15,7 @@ class CreateViewUsersPostsUpUlView extends Migration
     {
         \DB::statement("
     CREATE VIEW view_users_posts_up_ul AS
-        SELECT 
+        SELECT
             users.id AS id,
             posts.id AS post_id,
             user_profile.first_name AS first_name,
@@ -37,14 +37,14 @@ class CreateViewUsersPostsUpUlView extends Migration
             user_profile.user_id AS user_id
         FROM
             users
-        LEFT JOIN 
-            user_profile ON users.id::integer = user_profile.user_id::integer
+        LEFT JOIN
+            user_profile ON users.id = user_profile.user_id
 
-        LEFT JOIN 
-            user_lands ON users.id::integer = user_lands.user_id::integer
+        LEFT JOIN
+            user_lands ON users.id = user_lands.user_id
 
-        LEFT JOIN 
-            posts ON users.id::integer = posts.user_id::integer
+        LEFT JOIN
+            posts ON users.id = posts.user_id
         ");
     }
 
@@ -54,6 +54,6 @@ class CreateViewUsersPostsUpUlView extends Migration
      * @return void
      */
     public function down()
-    { 
+    {
     }
 }
